merge amd64 (basically dev) into main

+++ include/cpu/arch.h
@@ -14,9 +14,9 @@
  *  - The Raspberry Pi 4 port will provide an AArch64 implementation with the same API.
  */
 
-#include <types.h>
+#include <misc/types.h>
 
-typedef uint32 arch_irq_state_t;
+typedef uintptr arch_irq_state_t;
 
 /* Interrupt state management */
 void arch_disable_interrupts(void);
